smart-budget-agent/requirements.txt
firebase-admin==6.2.0
python-dotenv>=1.1.1,<2.0.0
google-auth-oauthlib>=1.2.2,<2.0.0
google-api-python-client>=2.176.0,<3.0.0
google-auth-httplib2>=0.2.0,<0.3.0
google-cloud-aiplatform==1.38.1
google-cloud-storage>=1.32.0,<3.0.0
pandas
numpy
scikit-learn
